requirements.txt
a2a-sdk[http-server]>=0.3.26
uvicorn[standard]>=0.30
httpx>=0.28
